requirements.txt
streamlit>=1.42,<2.0
pandas>=2.2,<3.0
requests>=2.31,<3.0
python-dotenv>=1.0,<2.0
openai>=1.65,<2.0
yfinance>=0.2.54,<0.3.0
